[SPIR-V] Check if the extf op is preceded by a bitcast op.
When native bf16 support is enabled, extf is not preceded by a bitcast op
(which is the case for bf16-to-gpu pass path, or non-native path),
and sometimes the operand to the extf may be coming from not an op,
rather an argument passed to a function, which may cause assert.
This check would circumvent that issue.

Also added the check for bitcast op handling.