SPIRV vector.mask lowering: use 64-bit type.
So far, most of the kernels use 16 or 32 vector sizes. But there is a
kernel that returns an i8 tensor and the vector size = 64. So, there
was an overflow when computing the mask (1 << valid_elements). This PR
updates the data type from 32-bit to 64-bit when computing the mask.
It also adds an assert to catch more easily this type of issue.